Change dirx diry to float

+++ headers/Spells.hpp
@@ -3,11 +3,11 @@
 
 class Spell : public MapObject {
     public:
-        Spell(int start_x, int start_y, int dir_x, int dir_y, int width, int height, bool noclip);
+        Spell(int start_x, int start_y, float dir_x, float dir_y, int width, int height, bool noclip);
         virtual void draw(int camera_x, int camera_y);
         virtual void move();
 
     protected:
-    int dir_x;
-    int dir_y;
+    float dir_x;
+    float dir_y;
 };
